pyproject.toml
[project]
name = "llm-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.1",
    "requests>=2.32.4",
    "zenml>=0.84.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
    "asyncio>=3.4.3",
    "pymongo>=4.6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
]
